requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.26
sqlparse==0.5.4
typing-extensions==4.13.2
asgiref==3.8.1
backports.zoneinfo==0.2.1
django==4.2.26
psycopg2-binary==2.9.10
sqlparse==0.5.4
typing-extensions==4.13.2
asgiref==3.11.0
Django==5.2
django-countries==8.2.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.21
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
typing_extensions==4.15.0
asgiref==3.11.0
Django==5.2
django-ckeditor==6.7.3
django-countries==8.2.0
django-js-asset==3.1.2
django-phonenumber-field==8.4.0
phonenumbers==9.0.21
pillow==12.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
typing_extensions==4.15.0
